refactor: logging to file only, show exception with console for client
